main_framework/requirements.txt
transformers
Pillow
tqdm
python-dotenv
openai